Create Appointments component
