+++ src/App.js
@@ -1,5 +1,6 @@
-import { BrowserRouter as Router } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Nav from './components/Nav/Nav';
+import AddAppointment from './components/AddAppointment/AddAppointment';
 import Appointment from './components/Appointment/Appointment';
 import './App.css';
 
@@ -8,8 +9,12 @@ function App() {
     <>
       <Router>
         <Nav />
+        <Routes>
+          <Route path="/appointments" exact element={<Appointment />} />
+        </Routes>
+        <Appointment />
       </Router>
-      <Appointment />
+      <AddAppointment />
     </>
   );
 }
