+++ src/components/AddAppointment/AddAppointment.css
@@ -1,0 +1,112 @@
+.add-appointment-modal {
+  background: #98bf0c;
+  height: 100vh;
+  position: fixed;
+  width: 100vw;
+}
+
+.show {
+  display: block;
+}
+
+.appointment-head {
+  padding: 40px;
+}
+
+.close-modal-button {
+  background: none;
+  border: none;
+  cursor: pointer;
+}
+
+.modal-close-icon {
+  color: #fff;
+  font-size: 35px;
+}
+
+.appointment-info {
+  padding: 40px;
+  color: #fff;
+  font-size: 20px;
+  font-family: cursive;
+}
+
+.add-appointment-section {
+  width: 70%;
+  margin: 0 auto;
+  padding: 40px 0;
+}
+
+.add-appointment-form {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 30px;
+}
+
+.doctors-date-time {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+  width: 40%;
+}
+
+.description-submit {
+  display: flex;
+  flex-direction: column;
+  width: 60%;
+  gap: 30px;
+  justify-content: flex-end;
+  align-items: center;
+}
+
+
+
+.select-doctors, .city-input, .date-input, .submit-appointment {
+  font-size: 18px;
+  padding: 7px;
+  border-radius: 8px;
+  color: #262525;
+  border: none;
+}
+
+.submit-appointment {
+  background: #585a54;
+  color: #fff;
+  cursor: pointer;
+}
+
+.city-input:focus, .date-input:focus {
+  outline: none;
+}
+
+.select-doctors {
+  width: 100%;
+}
+
+.city-input {
+  width: 100%;
+  height: 120px;
+}
+
+.submit-appointment {
+  width: max-content;
+}
+
+@media screen and (max-width: 768px) {
+  .add-appointment-form {
+    flex-direction: column;
+  }
+
+  .doctors-date-time {
+    width: 100%;
+  }
+
+  .description-submit {
+    width: 100%;
+  }
+
+  .select-doctors, .city-input, .date-input {
+    width: 100%;
+  }
+}
