+++ src/components/Appointment/Appointment.css
@@ -1,112 +1,33 @@
-.add-appointment-modal {
-  background: #98bf0c;
-  height: 100vh;
-  position: fixed;
-  width: 100vw;
+.appointments-container {
+  background: #fff;
 }
 
-.show {
-  display: block;
-}
-
-.appointment-head {
-  padding: 40px;
-}
-
-.close-modal-button {
-  background: none;
-  border: none;
-  cursor: pointer;
-}
-
-.modal-close-icon {
-  color: #fff;
-  font-size: 35px;
-}
-
-.appointment-info {
-  padding: 40px;
-  color: #fff;
-  font-size: 20px;
-  font-family: cursive;
-}
-
-.add-appointment-section {
-  width: 70%;
+.appointments-head {
+  text-align: center;
+  width: 90%;
   margin: 0 auto;
-  padding: 40px 0;
 }
 
-.add-appointment-form {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 30px;
+.appointments-heading {
+  color: grey;
+  font-family: cursive;
+  font-size: 30px;
 }
 
-.doctors-date-time {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  width: 40%;
+.my-appointments-main {
+  width: 90%;
+  margin: 0 auto;
 }
 
-.description-submit {
-  display: flex;
-  flex-direction: column;
-  width: 60%;
-  gap: 30px;
-  justify-content: flex-end;
-  align-items: center;
+.my-appointments-table {
+  margin: 0 auto;
 }
 
-
-
-.select-doctors, .city-input, .date-input, .submit-appointment {
-  font-size: 18px;
-  padding: 7px;
-  border-radius: 8px;
-  color: #262525;
+.delete-appointment-button {
   border: none;
-}
-
-.submit-appointment {
-  background: #585a54;
+  background: #cd1515;
   color: #fff;
+  border-radius: 3px;
+  padding: 5px;
   cursor: pointer;
-}
-
-.city-input:focus, .date-input:focus {
-  outline: none;
-}
-
-.select-doctors {
-  width: 100%;
-}
-
-.city-input {
-  width: 100%;
-  height: 120px;
-}
-
-.submit-appointment {
-  width: max-content;
-}
-
-@media screen and (max-width: 768px) {
-  .add-appointment-form {
-    flex-direction: column;
-  }
-
-  .doctors-date-time {
-    width: 100%;
-  }
-
-  .description-submit {
-    width: 100%;
-  }
-
-  .select-doctors, .city-input, .date-input {
-    width: 100%;
-  }
 }
